Make minor improvements and fixes
- Fix export for gif, jpeg, jpg, etc.
- Add logfile
- Allow pages to fail with warning
- Don't process recycle bin unless specifically enabled

+++ convert.py
@@ -5,11 +5,20 @@ import shutil
 import fitz
 import win32com.client as win32
 import pywintypes
+import re
 import traceback
 from xml.etree import ElementTree
 
 OUTPUT_DIR = os.path.join(os.path.expanduser('~'), "Desktop", "OneNoteExport")
 ASSETS_DIR = "assets"
+PROCESS_RECYCLE_BIN = False
+LOGFILE = 'onenote_to_markdown.log'
+
+def log(message):
+    print(message)
+    if LOGFILE is not None:
+        with open(LOGFILE, 'a') as lf:
+            lf.write("%s\n" % message)
 
 def safe_str(name):
     return  re.sub(r'[^.a-zA-Z0-9]', '_', name)
@@ -25,7 +34,7 @@ def extract_pdf_pictures(pdf_path, assets_path, page_name):
             pix = fitz.Pixmap(doc, xref)
             png_name = "%s_%s.png" % (page_name, str(img_num).zfill(3))
             png_path = os.path.join(assets_path, png_name)
-            print("Writing png: %s" % png_path)
+            log("Writing png: %s" % png_path)
             if pix.n < 5:
                 pix.save(png_path)
             else:
@@ -44,8 +53,7 @@ def fix_image_names(md_path, image_names):
         with open(tmp_path, 'w', encoding='utf-8') as f_tmp:
             body_md = f_md.read()
             for i,name in enumerate(image_names):
-                # TODO assumes all are png - maybe an issue?
-                body_md = body_md.replace("media/image%d.png" % (i+1), name)
+                body_md = re.sub("media\/image" + str(i+1) + "\.[a-zA-Z]+", name, body_md)
             f_tmp.write(body_md)
     shutil.move(tmp_path, md_path)
 
@@ -63,20 +71,25 @@ def handle_page(onenote, elem, path, i):
         os.remove(path_docx)
     if os.path.exists(path_pdf):
         os.remove(path_pdf)
-    # Create docx
-    onenote.Publish(elem.attrib['ID'], path_docx, win32.constants.pfWord, "")
-    # Convert docx to markdown
-    print("Generating markdown: %s" % path_md)
-    os.system('pandoc.exe -i %s -o %s -t markdown-simple_tables-multiline_tables-grid_tables --wrap=none' % (path_docx, path_md))
-    # Create pdf (for the picture assets)
-    onenote.Publish(elem.attrib['ID'], path_pdf, 3, "")
-    # Output picture assets to folder
-    image_names = extract_pdf_pictures(path_pdf, path_assets, safe_name)
-    # Replace image names in markdown file
-    fix_image_names(path_md, image_names)
+    try:
+        # Create docx
+        onenote.Publish(elem.attrib['ID'], path_docx, win32.constants.pfWord, "")
+        # Convert docx to markdown
+        log("Generating markdown: %s" % path_md)
+        os.system('pandoc.exe -i %s -o %s -t markdown-simple_tables-multiline_tables-grid_tables --wrap=none' % (path_docx, path_md))
+        # Create pdf (for the picture assets)
+        onenote.Publish(elem.attrib['ID'], path_pdf, 3, "")
+        # Output picture assets to folder
+        image_names = extract_pdf_pictures(path_pdf, path_assets, safe_name)
+        # Replace image names in markdown file
+        fix_image_names(path_md, image_names)
+    except pywintypes.com_error as e:
+        log("!!WARNING!! Page Failed: %s" % path_md)
     # Clean up docx, html
-    os.remove(path_docx)
-    os.remove(path_pdf)
+    if os.path.exists(path_docx):
+        os.remove(path_docx)
+    if os.path.exists(path_pdf):
+        os.remove(path_pdf)
 
 def handle_element(onenote, elem, path='', i=0):
     if elem.tag.endswith('Notebook'):
@@ -87,7 +100,7 @@ def handle_element(onenote, elem, path='', i=0):
         hier2 = onenote.GetHierarchy(elem.attrib['ID'], win32.constants.hsPages, "")
         for i,c2 in enumerate(ElementTree.fromstring(hier2)):
             handle_element(onenote, c2, os.path.join(path, safe_str(elem.attrib['name'])), i)
-    elif elem.tag.endswith('SectionGroup'):
+    elif elem.tag.endswith('SectionGroup') and (not elem.attrib['name'].startswith('OneNote_RecycleBin') or PROCESS_RECYCLE_BIN):
         hier2 = onenote.GetHierarchy(elem.attrib['ID'], win32.constants.hsSections, "")
         for i,c2 in enumerate(ElementTree.fromstring(hier2)):
             handle_element(onenote, c2, os.path.join(path, safe_str(elem.attrib['name'])), i)
@@ -106,4 +119,4 @@ if __name__ == "__main__":
 
     except pywintypes.com_error as e:
         traceback.print_exc()
-        print("!!!Error!!! Hint: Make sure OneNote is open first.")
+        log("!!!Error!!! Hint: Make sure OneNote is open first.")
